delete the login page and password insert
