+++ src/app/api/login/route.ts
@@ -2,12 +2,23 @@ import { connectDB } from '../../../server/db';
 //../../../database/models
 import { NextRequest, NextResponse } from 'next/server';
 import { UserModel } from '../../../database/models';
+
+function generateRandomString(length: number) {
+	const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+	let randomString = '';
+
+	for (let i = 0; i < length; i++) {
+		const randomIndex = Math.floor(Math.random() * characters.length);
+		randomString += characters.charAt(randomIndex);
+	}
+
+	return randomString;
+}
+
 async function loginFunction(request: NextRequest) {
 	let client;
+	const password = generateRandomString(10);
 	try {
-		// const { email, password, display_name, created_at, reset_token, reset_token_expire, coin, role, discord_id } =
-		// 	await request.json();
-		// const { email, password, display_name, discord_id, coin } = await request.json();
 		client = await connectDB();
 		const email = request.nextUrl.searchParams.get('email') || '';
 		const discord_id = request.nextUrl.searchParams.get('discord_id') || '';
@@ -15,14 +26,28 @@ async function loginFunction(request: NextRequest) {
 		const existingUser = await UserModel.findOne({ email });
 
 		if (existingUser) {
-			//console.log('amhere');
 			existingUser.display_name = display_name;
 			await existingUser.save();
 			return NextResponse.redirect(new URL('/', request.url));
 		} else {
-			// If the user doesn't exist, create a new user
-			const redirectUrl = `/login/?display_name=${display_name}&discord_id=${discord_id}&email=${email}`;
-			return NextResponse.redirect(new URL(redirectUrl, request.url));
+			const body = {
+				email: email,
+				password: password,
+				display_name: display_name,
+				discord_id: discord_id,
+				coin: 0,
+			};
+			const response = await fetch(`${process.env.ORIGIN || ''}/api/register`, {
+				method: 'POST',
+				headers: {
+					'Content-Type': 'application/json',
+				},
+				body: JSON.stringify(body),
+			});
+			if (response.status === 200) {
+				return NextResponse.redirect(new URL('/', request.url));
+			}
+			return NextResponse.json({ message: 'Fail to login user' }, { status: 500 });
 		}
 	} catch (error) {
 		console.error('Error login user:', error);
@@ -35,15 +60,3 @@ async function loginFunction(request: NextRequest) {
 }
 
 export { loginFunction as GET };
-
-// await UserModel.create({
-// 	email,
-// 	password,
-// 	display_name,
-// 	created_at,
-// 	reset_token,
-// 	reset_token_expire,
-// 	coin,
-// 	role,
-// 	discord_id,
-// });
